refactor(introduce-strategy-pattern): 전략 패턴 적용
덱에 들어갈 카드를 생성하는 로직에 전략 패턴을 적용했습니다.
실제 로직에 사용될 구현체에는 카드 52장을 생성하고 셔플하는 로직을 넣었고, 테스트를 위한 구현체에는 테스트에 필요한만큼만 카드를 순서대로 넣는 로직을 넣었습니다.
