+++ src/main/java/service/CardGenerator.java
@@ -1,0 +1,9 @@
+package service;
+
+import domain.card.Card;
+import java.util.List;
+
+public interface CardGenerator {
+
+    List<Card> generate();
+}
