+++ src/test/java/domain/card/DeckTest.java
@@ -1,6 +1,6 @@
 package domain.card;
 
-import static message.ErrorMessage.*;
+import static message.ErrorMessage.DECK_CAN_NOT_DUPLICATED;
 
 import domain.enums.Rank;
 import domain.enums.Suit;
@@ -9,13 +9,16 @@ import java.util.Set;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import service.CardGenerator;
+import service.ShuffledCardGenerator;
 
 public class DeckTest {
 
     @DisplayName("52장의 서로 다른 카드가 정상 생성된다.")
     @Test
     void _52장의_서로_다른_카드가_정상_생성된다() {
-        Deck deck = new Deck(CardGenerator.generateCards());
+        CardGenerator cardGenerator = new ShuffledCardGenerator();
+        Deck deck = new Deck(cardGenerator.generate());
 
         Set<Card> distinctCards = Set.copyOf(deck.getCards());
 
@@ -26,7 +29,8 @@ public class DeckTest {
     @DisplayName("중복된 카드가 들어있을 경우 예외가 발생한다.")
     @Test
     void 중복_카드_예외_발생_한다() {
-        List<Card> cards = CardGenerator.generateCards();
+        CardGenerator cardGenerator = new ShuffledCardGenerator();
+        List<Card> cards = cardGenerator.generate();
         cards.add(new Card(Rank.ACE, Suit.SPADE));
 
         Assertions.assertThatThrownBy(() -> new Deck(cards))
